Refactored and extracted styling and components
